Use _ENV to export module values
The require function inserts all global values from the module it
imports in the _G table. Because of that, you need in principle to set
only local variables in the module's code in order to not pollute the
global namespace. So, what's the purpose of the "local" keyword if it
needs to be used for all declarations?

I don't like this behaviour. It would be better if "require" assume
every non-local variable to be exported. More or less like this:

function require(module)
    local env = {}
    loadfile(module, "bt", env)()
    return env
end

Since require doesn't act like this, you can emulate this behaviour by
setting _ENV to a local table, like this:

local M = {}; _ENV = M